+++ app/(user)/project/donation/[id]/components/ProjectSidebar.tsx
@@ -1,18 +1,18 @@
-import type { Project } from "@/types/project/project";
+// import type { Project } from "@/types/project/project";
 
-import ProjectSidebarDonation from "./ProjectSidebarDonation";
-import ProjectSidebarPanel from "./ProjectSidebarPanel";
+// import ProjectSidebarDonation from "./ProjectSidebarDonation";
+// import ProjectSidebarPanel from "./ProjectSidebarPanel";
 
-interface Props {
-  project: Project;
-}
+// interface Props {
+//   project: Project;
+// }
 
-export default function ProjectSidebar({ project }: Props) {
-  return (
-    <div className="space-y-6">
-      {/* <ProjectSidebarDonation project={project} />
+// export default function ProjectSidebar({ project }: Props) {
+//   return (
+//     <div className="space-y-6">
+//       {/* <ProjectSidebarDonation project={project} />
 
-      <ProjectSidebarPanel creator={project.creator} stats={project.stats} /> */}
-    </div>
-  );
-}
+//       <ProjectSidebarPanel creator={project.creator} stats={project.stats} /> */}
+//     </div>
+//   );
+// }
