+++ app/toss/page.tsx
@@ -1,23 +1,27 @@
 "use client";
 
+import { Suspense, useEffect, useState } from "react";
+import { useSearchParams } from "next/navigation";
 import {
   loadTossPayments,
   TossPaymentsWidgets,
 } from "@tosspayments/tosspayments-sdk";
-import { useEffect, useState } from "react";
-import { useSearchParams } from "next/navigation";
 import "./App.css";
 
-function generateRandomString() {
-  if (typeof window !== "undefined") {
-    return window.btoa(Math.random().toString()).slice(0, 20);
-  }
-  return ""; // 서버 환경일 경우 기본값 반환
+export default function TossPage() {
+  return (
+    <Suspense fallback={<p>로딩 중...</p>}>
+      <InnerTossPage />
+    </Suspense>
+  );
 }
 
-// TODO: clientKey는 개발자센터의 결제위젯 연동 키 > 클라이언트 키로 바꾸세요.
-// TODO: 구매자의 고유 아이디를 불러와서 customerKey로 설정하세요. 이메일・전화번호와 같이 유추가 가능한 값은 안전하지 않습니다.
-// @docs https://docs.tosspayments.com/sdk/v2/js#토스페이먼츠-초기화
+function generateRandomString() {
+  return typeof window !== "undefined"
+    ? window.btoa(Math.random().toString()).slice(0, 20)
+    : "";
+}
+
 const clientKey = "test_gck_docs_Ovk5rk1EwkEbP0W43n07xlzm";
 const customerKey = generateRandomString();
 
@@ -26,10 +30,11 @@ interface Amount {
   value: number;
 }
 
-export default function CheckoutPage() {
+function InnerTossPage() {
   const params = useSearchParams();
   const orderId = params.get("orderId")!;
   const initialAmount = Number(params.get("amount")!);
+
   const [amount, setAmount] = useState<Amount>({
     currency: "KRW",
     value: initialAmount,
@@ -38,63 +43,31 @@ export default function CheckoutPage() {
   const [widgets, setWidgets] = useState<TossPaymentsWidgets | null>(null);
 
   useEffect(() => {
-    async function fetchPaymentWidgets() {
+    (async () => {
       try {
-        // ------  SDK 초기화 ------
-        // @docs https://docs.tosspayments.com/sdk/v2/js#토스페이먼츠-초기화
         const tossPayments = await loadTossPayments(clientKey);
-
-        // 회원 결제
-        // @docs https://docs.tosspayments.com/sdk/v2/js#tosspaymentswidgets
-        const widgets = tossPayments.widgets({
-          customerKey,
-        });
-        // 비회원 결제
-        // const widgets = tossPayments.widgets({ customerKey: ANONYMOUS });
-
+        const widgets = tossPayments.widgets({ customerKey });
         setWidgets(widgets);
       } catch (error) {
-        console.error("Error fetching payment widget:", error);
+        console.error(error);
       }
-    }
-
-    fetchPaymentWidgets();
-    // eslint-disable-next-line
-  }, [clientKey, customerKey]);
+    })();
+  }, []);
 
   useEffect(() => {
-    async function renderPaymentWidgets() {
-      if (widgets == null) {
-        return;
-      }
-
-      // ------  주문서의 결제 금액 설정 ------
-      // TODO: 위젯의 결제금액을 결제하려는 금액으로 초기화하세요.
-      // TODO: renderPaymentMethods, renderAgreement, requestPayment 보다 반드시 선행되어야 합니다.
+    (async () => {
+      if (!widgets) return;
       await widgets.setAmount(amount);
-
-      // ------  결제 UI 렌더링 ------
-      // @docs https://docs.tosspayments.com/sdk/v2/js#widgetsrenderpaymentmethods
       await widgets.renderPaymentMethods({
         selector: "#payment-method",
-        // 렌더링하고 싶은 결제 UI의 variantKey
-        // 결제 수단 및 스타일이 다른 멀티 UI를 직접 만들고 싶다면 계약이 필요해요.
-        // @docs https://docs.tosspayments.com/guides/v2/payment-widget/admin#새로운-결제-ui-추가하기
         variantKey: "DEFAULT",
       });
-
-      // ------  이용약관 UI 렌더링 ------
-      // @docs https://docs.tosspayments.com/reference/widget-sdk#renderagreement선택자-옵션
       await widgets.renderAgreement({
         selector: "#agreement",
         variantKey: "AGREEMENT",
       });
-
       setReady(true);
-    }
-
-    renderPaymentWidgets();
-    // eslint-disable-next-line
+    })();
   }, [widgets]);
 
   const updateAmount = async (amount: Amount) => {
@@ -105,11 +78,8 @@ export default function CheckoutPage() {
   return (
     <div className="wrapper">
       <div className="box_section">
-        {/* 결제 UI */}
         <div id="payment-method" />
-        {/* 이용약관 UI */}
         <div id="agreement" />
-        {/* 쿠폰 체크박스 */}
         <div style={{ paddingLeft: "24px" }}>
           <div className="checkable typography--p">
             <label
@@ -120,10 +90,7 @@ export default function CheckoutPage() {
                 id="coupon-box"
                 className="checkable__input"
                 type="checkbox"
-                aria-checked="true"
                 disabled={!ready}
-                // ------  주문서의 결제 금액이 변경되었을 경우 결제 금액 업데이트 ------
-                // @docs https://docs.tosspayments.com/sdk/v2/js#widgetssetamount
                 onChange={async event => {
                   await updateAmount({
                     currency: amount.currency,
@@ -137,18 +104,14 @@ export default function CheckoutPage() {
             </label>
           </div>
         </div>
-
-        {/* 결제하기 버튼 */}
         <button
           className="button"
           style={{ marginTop: "30px" }}
           disabled={!ready}
-          // ------ '결제하기' 버튼 누르면 결제창 띄우기 ------
-          // @docs https://docs.tosspayments.com/sdk/v2/js#widgetsrequestpayment
           onClick={async () => {
             try {
               await widgets!.requestPayment({
-                orderId: orderId,
+                orderId,
                 orderName: "토스 티셔츠 외 2건",
                 successUrl: window.location.origin + "/toss/success",
                 failUrl: window.location.origin + "/toss/fail",
@@ -157,7 +120,6 @@ export default function CheckoutPage() {
                 customerMobilePhone: "01012341234",
               });
             } catch (error) {
-              // 에러 처리하기
               console.error(error);
             }
           }}
